requirements.txt
# Core data
numpy>=1.23,<3
pandas>=1.5,<3

# ML
scikit-learn>=1.2
joblib>=1.3

# Experiment tracking
mlflow>=2.9

# Database
psycopg2-binary>=2.9

# API / Deployment
fastapi>=0.110
uvicorn>=0.23
flask>=3.0
streamlit>=1.30
django>=5.2
djangorestframework>=3.15

# NLP (LLM)
transformers>=4.40
sentencepiece>=0.2.0
tokenizers>=0.19

# Optional visualization
matplotlib>=3.7